Add native/gem5 compilation option; Add support for neon; Add scalar/neon runtime option.

+++ csr_formatter.h
@@ -7,7 +7,7 @@
 using namespace std;
 
 struct CSR {
-	vector<double> val;
+	vector<FLOAT> val;
 	vector<int>	col_ind;
 	vector<int> row_ptr;
 	int row;
